+++ fast_rx/test/rx/composite/rx_composite_test.dart
@@ -1,0 +1,36 @@
+import 'package:fast_rx_test/fast_rx_test.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'composite.dart';
+
+void main() {
+  final shouldNotify = <RxTest<Composite>>[
+    RxTest(() => Composite(), (rx) => rx.a.value = 2),
+    RxTest(
+      () => Composite(),
+      (rx) => rx.run(() {
+        rx.a.value = 2;
+        rx.b.value = 3;
+        rx.c.value = 4;
+      }),
+    ),
+  ];
+
+  final shouldRegister = <RxTest<Composite>>[
+    RxTest(() => Composite(), (rx) => rx.a.value),
+  ];
+
+  test('RxComposite notifications', () {
+    expectRxNotification(
+      shouldNotify: shouldNotify,
+      shouldNotNotify: shouldRegister,
+    );
+  });
+
+  test('RxComposite registration', () {
+    expectRxRegistration(
+      shouldRegister: shouldRegister,
+      shouldNotRegister: shouldNotify,
+    );
+  });
+}