+++ fast_rx/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.3.0
+- Adds `RxComposite` to allow for composition of `Rx` objects
+- Adds `RxCompositeList`, `RxCompositeSet`, and `RxCompositeMap`
+- Adds `composite` extension on `RxIterable<Rx>`s to create composite versions of them
+- `Rx.value` field moved to `RxValue.value`
+- `RxObject` now extends `RxValue`
+- `RxValue ==` now behaves as expected
+- Removes `Rx.close`
+- A bunch of refactoring
+
 ## 0.2.10
 - Fixes an edge case where registration could occur outside of a `FastBuilder` builder